requirements.txt
Flask>=2.3.0
flask-sqlalchemy>=3.0.0
flask-login>=0.6.0
flask-migrate>=4.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.31.0
psycopg2-binary>=2.9.0
spotipy>=2.23.0
numpy>=1.24.0
werkzeug>=2.3.0
